+++ lib/generators/paper_trail/install_generator.rb
@@ -1,5 +1,4 @@
 require 'rails/generators'
-require 'rails/generators/migration'
 require 'rails/generators/active_record'
 
 module PaperTrail
@@ -13,7 +12,7 @@ module PaperTrail
 
     def create_migration_file
       add_paper_trail_migration('create_versions')
-      add_paper_trail_migration('add_object_changes_column_to_versions') if options.with_changes?
+      add_paper_trail_migration('add_object_changes_to_versions') if options.with_changes?
       add_paper_trail_migration('create_version_associations')
       add_paper_trail_migration('add_transaction_id_column_to_versions')
     end