+++ lib/generators/paper_trail/templates/add_object_changes_column_to_versions.rb
@@ -1,9 +1,0 @@
-class AddObjectChangesColumnToVersions < ActiveRecord::Migration
-  def self.up
-    add_column :versions, :object_changes, :text
-  end
-
-  def self.down
-    remove_column :versions, :object_changes
-  end
-end