+++ lib/generators/paper_trail/templates/add_object_changes_to_versions.rb
@@ -1,0 +1,5 @@
+class AddObjectChangesToVersions < ActiveRecord::Migration
+  def change
+    add_column :versions, :object_changes, :text
+  end
+end