+++ lib/paper_trail.rb
@@ -34,12 +34,12 @@ module PaperTrail
 
   # Sets whether PaperTrail is enabled or disabled for this model in the current request.
   def self.enabled_for_model(model, value)
-    paper_trail_store[:"request_enabled_for_#{model}"] = value
+    paper_trail_store[:"enabled_for_#{model}"] = value
   end
 
   # Returns `true` if PaperTrail is enabled for this model in the current request, `false` otherwise.
   def self.enabled_for_model?(model)
-    !!paper_trail_store.fetch(:"request_enabled_for_#{model}", true)
+    !!paper_trail_store.fetch(:"enabled_for_#{model}", true)
   end
 
   # Set the field which records when a version was created.