+++ spec/paper_trail_spec.rb
@@ -5,12 +5,20 @@ describe "PaperTrail RSpec Helper" do
     it 'should have versioning off by default' do
       ::PaperTrail.should_not be_enabled
     end
-    it 'should turn versioning on in a with_versioning block' do
+    it 'should turn versioning on in a `with_versioning` block' do
       ::PaperTrail.should_not be_enabled
       with_versioning do
         ::PaperTrail.should be_enabled
       end
       ::PaperTrail.should_not be_enabled
+    end
+
+    context "error within `with_versioning` block" do
+      it "should revert the value of `PaperTrail.enabled?` to it's previous state" do
+        ::PaperTrail.should_not be_enabled
+        expect { with_versioning { raise } }.to raise_error
+        ::PaperTrail.should_not be_enabled
+      end
     end
   end
 
@@ -24,6 +32,19 @@ describe "PaperTrail RSpec Helper" do
         ::PaperTrail.should be_enabled
       end
       ::PaperTrail.should be_enabled
+    end
+  end
+
+  context '`with_versioning` block at class level' do
+    it { ::PaperTrail.should_not be_enabled }
+
+    with_versioning do
+      it 'should have versioning on by default' do
+        ::PaperTrail.should be_enabled
+      end
+    end
+    it 'should not leak the `enabled?` state into successive tests' do
+      ::PaperTrail.should_not be_enabled
     end
   end